backend/requirements.txt
fastapi>=0.115
uvicorn[standard]>=0.30
pydantic>=2.8
pydantic-settings>=2.4
sqlalchemy>=2.0
alembic>=1.13
psycopg2-binary>=2.9
python-jose[cryptography]>=3.3
passlib[bcrypt]>=1.7
python-multipart>=0.0.9
aiofiles>=24.0
httpx>=0.27

# LangChain / RAG
langchain>=1.0
langchain-community>=0.3
langchain-groq>=1.0
langchain-chroma>=1.0
langchain-text-splitters>=0.3
langchain-core>=1.0
chromadb>=1.0
sentence-transformers>=3.0

# ML
ultralytics>=8.2
opencv-python-headless>=4.10
pillow>=10.0
numpy>=1.26
pandas>=2.2
scikit-learn>=1.5
scipy>=1.14

# Utilities
python-dotenv>=1.0
pypdf>=4.0
